requirements.txt
python-dotenv
nltk
pandas
requests
python-Levenshtein
